backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
python-dateutil
psutil
tzdata
